+++ tests/testthat/test_as_task_fcst.R
@@ -65,7 +65,7 @@ test_that("as_task conversion keyed tsibble", {
   expect_class(task, "TaskFcst")
   expect_equal(task$col_roles$key, "id")
   expect_true("keys" %in% task$properties)
-  expect_factor(task$data(cols = "id")$id)
+  expect_character(task$data(cols = "id")$id)
 })
 
 test_that("as_task_fcst assertions", {